Add more generic scripts to setup instances, run performance tests, and summarize results.

+++ distribution/scripts/jmeter/jmeter-server-start.sh
@@ -1,4 +1,4 @@
-#!/bin/bash
+#!/bin/bash -e
 # Copyright 2017 WSO2 Inc. (http://wso2.org)
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -18,24 +18,73 @@
 # ----------------------------------------------------------------------------
 
 script_dir=$(dirname "$0")
+
+jar_name=ApacheJMeter.jar
+jmeter_hostname=""
+jmeter_installation_dir=""
+jmeter_heap_size=""
+
+function usage() {
+    echo ""
+    echo "Usage: "
+    echo "$0 -n <jmeter_hostname> -i <jmeter_installation_dir> [-h] -- [jmeter_flags]"
+    echo ""
+    echo "-n: The JMeter hostname."
+    echo "-i: The JMeter installation directory."
+    echo "-m: The heap memory size of JMeter."
+    echo "-h: Display this help and exit."
+    echo ""
+}
+
+while getopts "n:i:m:h" opts; do
+    case $opts in
+    n)
+        jmeter_hostname=${OPTARG}
+        ;;
+    i)
+        jmeter_installation_dir=${OPTARG}
+        ;;
+    m)
+        jmeter_heap_size=${OPTARG}
+        ;;
+    h)
+        usage
+        exit 0
+        ;;
+    \?)
+        usage
+        exit 1
+        ;;
+    esac
+done
+shift "$((OPTIND - 1))"
+
+jmeter_flags="$@"
+
 JMETER_HOME=""
-for dir in $HOME/apache-jmeter-*; do
+for dir in $jmeter_installation_dir/apache-jmeter-*; do
     [ -d "${dir}" ] && JMETER_HOME="${dir}" && break
 done
 
-jar_name=ApacheJMeter.jar
-jmeter_host=$1
+if [[ -z $jmeter_hostname ]]; then
+    echo "Please provide a hostname for JMeter server."
+    exit 1
+fi
 
-if pgrep -f "$jar_name" > /dev/null; then
+if [[ ! -d $JMETER_HOME ]]; then
+    echo "Could not find JMeter home directory."
+    exit 1
+fi
+
+if pgrep -f "$jar_name" >/dev/null; then
     echo "Stopping JMeter Server"
     pkill -f $jar_name
 fi
 
 echo "Waiting for JMeter Server to stop"
 
-while true
-do
-    if ! pgrep -f "$jar_name" > /dev/null; then
+while true; do
+    if ! pgrep -f "$jar_name" >/dev/null; then
         echo "JMeter Server stopped"
         break
     else
@@ -50,11 +99,22 @@ if [[ -f $gc_log_file ]]; then
     mv $gc_log_file /tmp/
 fi
 
-export JVM_ARGS="-Xms4g -Xmx4g -XX:+PrintGC -XX:+PrintGCDetails -XX:+PrintGCDateStamps -Xloggc:$gc_log_file"
-export RMI_HOST_DEF=-Djava.rmi.server.hostname=$jmeter_host
+if [[ -z $jmeter_heap_size ]]; then
+    jmeter_heap_size="4g"
+fi
+
+export JVM_ARGS="-Xms$jmeter_heap_size -Xmx$jmeter_heap_size -XX:+PrintGC -XX:+PrintGCDetails -XX:+PrintGCDateStamps -Xloggc:$gc_log_file $jmeter_flags"
+export RMI_HOST_DEF=-Djava.rmi.server.hostname=$jmeter_hostname
 
 echo "Starting JMeter Server"
-nohup $JMETER_HOME/bin/jmeter-server > server.out 2>&1 &
+nohup $JMETER_HOME/bin/jmeter-server >server.out 2>&1 &
 
 # Sleep for 10 seconds and make sure the JMeter server is ready to run the tests
 sleep 10
+
+if pgrep -f "$jar_name" >/dev/null; then
+    echo "Started JMeter server."
+else
+    echo "JMeter Server has not started!"
+    exit 1
+fi
